Fix nulls bug in geo locations insertion

+++ DB/structure/sql/samples/create_table_samples.sql
@@ -34,7 +34,7 @@ create table samples (
 	sra_study text,
 	serotype text,
 
-	geo_location_id bigint,
+	geo_location_id integer,
 	assay_type text,
 	avg_spot_length double precision,
 	bases bigint,
